requirements.txt
openai
python-dotenv
streamlit
streamlit-option-menu
